Pull JuliaInfo out of core
